Updated pkg, added a 4th input box on dashboard for independent rpc status from in game status, created a script to save input that input and set it to state in discordHelper
V1 of independent RPC status from the in game status. I don't think I'm requesting the string correctly which is why its not setting it in discordHelper.Js.  Some comments are in there to help me remember the premise as I have ADHD. As well as some commands to build, or create a temp run, and remember some old code I've used and errors I've got.